remaned K2Raster files changed RavenJoin so it supports long k2-raster trees
